Can retrieve a list of the user’s desired repos
This way we can list the user’s orgs that they contribute to, and then from there, select the repos that they are interested in getting stats on.

Alternatively you can get a list of all repos that a user contributes to, but IMO this would be a better way to narrow down the desired repos.