Creates the Query Types for the Schema
Note that I started creating a sub directory for query types.  This is so that we don’t have all of the queries, for all of the models, all inside of one large file.

This encapsulate’s each model’s query logic inside of its own file, and then includes it in the main “app/graphql/types/query_type.rb” file.

This file is the one that is wired up to the schema, through the query root entry point.

I derived this implementation pattern from the HasRecommendedItems section of:

http://graphql-ruby.org/fields/resolvers.html