Implements the initial GraphQL server in Ruby (#39)
* Adds the GraphQL gems

* Default files from GraphQL generator

Install the current recommended default files and structure through this command:

rails generate graphql:install

See:

http://graphql-ruby.org/schema/generators#graphqlinstall

* Declares types for two models

See:

http://graphql-ruby.org/getting_started#declare-types

* Creates the Query Types for the Schema

Note that I started creating a sub directory for query types.  This is so that we don’t have all of the queries, for all of the models, all inside of one large file.

This encapsulate’s each model’s query logic inside of its own file, and then includes it in the main “app/graphql/types/query_type.rb” file.

This file is the one that is wired up to the schema, through the query root entry point.

I derived this implementation pattern from the HasRecommendedItems section of:

http://graphql-ruby.org/fields/resolvers.html

* Initial query type implementation for Statistic & GithubUser

* Renames the GraphQL schema

* Creates an extensible, organized, encapsulated query file structure

Now queries can be organized by the model (or any common theme) in which they fall under.  And each query resides in its own file.

The queries for a given model (or theme) role up into a manifest of queries.  This identifies all of the queries currently available for the given model (or theme).

Each model’s manifest role’s up into the root query, that is passed to the schema.  Which is the expectation and pattern set by GraphQL and the GraphQL gem.

* Favors new manifest based implementation over this one

* Initial queries for Statistic, GithubUser & User

* Ensures that GraphQL responds with an Integer not a string

* Implements initial Mutation pattern

This commit:

- implements the create and update mutations for Statistic
- sets the basic pattern for mutations, re: file/code organization
- uses an Input Object for the mutations

see: http://graphql-ruby.org/mutations/mutation_classes.html

see: http://graphql-ruby.org/type_definitions/input_objects.html#customizing-input-objects

* Makes timestamps returnable and inheritable

* Accounts for a Statistic not being found during update

* Spec samples for queries, mutations, & requests

* Remaining query & mutation spec coverage

* RuboCop fixes

* Required Yard doc additions per RuboCop

* Rake task export graphql schema (#40)

* adds a small task and test to update schema

* adds graphql schema

* tabs

* formatting