A user will need an organization to pull GitHub data
However, we cannot make this a direct Rails association, b/c we cannot validate for the presence of it when the user is being created.

Reason being, we are using OAuth to initially create the user, and at that point of the user being created, they will not yet have an GitHub organization associated with their record.