Column for tracking a statistic’s assignees
There are two ways a GithubUser can be associated with a Statistic.  Either they

- created the stat
- are assigned to the stat

Also, more than one user can be assigned to a given stat.

We need to be able to query against each of these cases.

By adding this column, as a type array, it allows us to query any of our MVP cases, all in one performant query (no N + 1’s, or creating more unnecessary database tables).