Creates one:many association b/w Stat/Repo
Everytime we create/update a Statistic, we always assign the repository.  We can do this b/c Repository is required information the GitHub’s API response.

So we can count on this always being present.

We will need this association in order to eager load a given stat’s organization and repository, for performant querying.