Fixes bug when creating a GithubUser from UpdateOrCreate class (#57)
* Adds service object to fetch a user’s GitHub account info

* Removes bug from UpdateOrCreate class

Sometimes the issue’s main user will be our user, and sometimes the issue’s main user will be some other user (and in this case, our user would be an assignee).

This was creating GithubUser records for the issue’s main user, regardless, which was the bug.

* Creates the GithubUser, based on our user

Now we are creating the GithubUser strictly based on our initialized user (which is always the case).  We are also saving a number of db checks by only checking for the user’s presence when Persist is initialized, not every single time UpdateOrCreate is iterated upon.