Optimize GraphQL for fetching statistics that meet all of the MVP requirements (#56)
* Favors conventional GraphQL naming patterns

Using query instead of root query
Using mutation instead of root mutation
Using StatisticInput instead of StatisticAttributes

* Column for tracking a statistic’s assignees

There are two ways a GithubUser can be associated with a Statistic.  Either they

- created the stat
- are assigned to the stat

Also, more than one user can be assigned to a given stat.

We need to be able to query against each of these cases.

By adding this column, as a type array, it allows us to query any of our MVP cases, all in one performant query (no N + 1’s, or creating more unnecessary database tables).

* Adds/updates a Statistic’s assignees

When we are creating and updating Statistic records, we need to also track/update who the issue’s assignees are.

We need this to be able to meet our MVP query goals.

* Creates one:many association b/w Stat/Repo

Everytime we create/update a Statistic, we always assign the repository.  We can do this b/c Repository is required information the GitHub’s API response.

So we can count on this always being present.

We will need this association in order to eager load a given stat’s organization and repository, for performant querying.

* Establishes all of the eligible enum types for FE querying

When the FE is making GraphQL queries, this creates the specific choices we will need to meet the MVP goal queries.

For example, being able to do things like this:

  statistics(
    type: [PR],
    state: [OPEN, MERGED],
    ownershipType: ASSIGNED,
    githubUserId: 14881910,
    datetimeType: UPDATED_AFTER,
…

* Queries and scope needed to get desired Statistics

With all of these reusable scopes and queries at the model layer, we can meet all of the needed MVP query cases.

It also leverages eager loading so the FE can get the repo and org info for each stat, without N + 1 queries.  For example,

  statistics(
    type: [PR],
    state: [OPEN, CLOSED, MERGED],
    ownershipType: ASSIGNED,
    githubUserId: 14881910,
    datetimeType: UPDATED_AFTER,
    datetime: "2018-10-06T20:31:41Z"
  ) {
    sourceType
    state
…
    repository {
      name
      url
    }
    organization {
      name
      url
    }
  }

* StatisticQuery can now be used for all MVP cases

For example, answers all of these questions with these queries:

## Issues ##
# How many issues did they create?
query {
  statistics(
    ownershipType: CREATED,
    githubUserId: 14881910,
    type: [ISSUE],
    state: [OPEN, CLOSED],
    datetimeType: CREATED_AFTER,
    datetime: "2018-10-06T20:31:41Z"
  ) {
    id
    source_id
    source_type
    source
    state
    repository {
      name
      url
    }
    organization {
      name
      url
    }
    url
    title
    source_created_at
    source_updated_at
    source_closed_at
  }
}

# How many issues did they work?
  statistics(
    ownershipType: ASSIGNED,
    githubUserId: 14881910,
    type: [ISSUE],
    state: [OPEN, CLOSED],
    datetimeType: UPDATED_AFTER,
    datetime: "2018-10-06T20:31:41Z"
  )

# How many issues did they close?
  statistics(
    ownershipType: ASSIGNED,
    githubUserId: 14881910,
    type: [ISSUE],
    state: [CLOSED],
    datetimeType: CLOSED_AFTER,
    datetime: "2018-10-06T20:31:41Z"
  )

## Pull Requests ##
# How many PRs did they create?
  statistics(
    ownershipType: CREATED,
    githubUserId: 14881910,
    type: [PR],
    state: [OPEN, MERGED],
    datetimeType: CREATED_AFTER,
    datetime: "2018-10-06T20:31:41Z"
  )

# How many PRs did they work?
  statistics(
    ownershipType: CREATED,
    githubUserId: 14881910,
    type: [PR],
    state: [OPEN, MERGED],
    datetimeType: UPDATED_AFTER,
    datetime: "2018-10-06T20:31:41Z"
  )

# How many PRs did they close?
  statistics(
    ownershipType: CREATED,
    githubUserId: 14881910,
    type: [PR],
    state: [MERGED],
    datetimeType: CLOSED_AFTER,
    datetime: "2018-10-06T20:31:41Z"
  )

# What are the links to each one of the above?

* FE can now get a GithubUser and all their stats, with one id

Now the FE can do a query that fetches a GithubUser and Statistics, all based on one ID, the github_id.

It would be something like:

query {
	githubUser(githubId:14881910) {
    githubLogin
  }
  statistics(
    type: [PR],
    state: [OPEN, CLOSED, MERGED],
    ownershipType: ASSIGNED,
    githubUserId: 14881910,
    datetimeType: UPDATED_AFTER,
    datetime: "2018-10-06T20:31:41Z"
  )
…