User can remove an item from their week in review (#80)
* Mutation to delete an Accomplishment from a WIR

* FE mutation to delete an accomplishment

* Removes clicked stat from week in review

Each stat can be clicked to remove it from that week’s “week in review”.  When clicked, the user is prompted to confirm the removal.  If they say yes, the stat is removed from the week in review.

This is done by deleting the associated Accomplishment record.  The associated Statistic record is not impacted.