Query can now return stats scoped to the week
The new for_week arg is set to false by default.

If set to true, it will scope the resulting statistics to fall in the week associated with the passed date time (i.e. within the Monday —> Sunday of that date time).

When for_week is false, it sets the “since” date time to the passed date time, and the “up through” date time to the current date time.