Week in review page is dynamic for selected week
For the current user, the Week In Review page will now displays:

- one week’s worth of statistics
- statistics dynamically, meaning it updates itself as the week progress, pulling in any new updates in the database
- the statistics for the selected week only
- the week’s start and end dates, that the statistics fall within
- the associated employee, meaning the current user

It does not create a `WeekInReview` db record, and the associated `Accomplishment` db records, as it used to.

This will not happen as part of a workflow where the user decided to “begin the week in review submittal process”, in upcoming stories.