Week in review stats by week for current user (#86)
* Week in review page is dynamic for selected week

For the current user, the Week In Review page will now displays:

- one week’s worth of statistics
- statistics dynamically, meaning it updates itself as the week progress, pulling in any new updates in the database
- the statistics for the selected week only
- the week’s start and end dates, that the statistics fall within
- the associated employee, meaning the current user

It does not create a `WeekInReview` db record, and the associated `Accomplishment` db records, as it used to.

This will not happen as part of a workflow where the user decided to “begin the week in review submittal process”, in upcoming stories.

* Dedicated class to handle start/end date/time logic

* Statistics query can now return stats scoped to the week

The new for_week arg is set to false by default.

If set to true, it will scope the resulting statistics to fall in the week associated with the passed date time (i.e. within the Monday —> Sunday of that date time).

When for_week is false, it sets the “since” date time to the passed date time, and the “up through” date time to the current date time.